+++ back-end/config/routes.rb
@@ -44,6 +44,8 @@ Rails.application.routes.draw do
       post :generate_modeled_image, on: :member, controller: :modeled_outfit_images, action: :create
     end
     resources :outfit_folders, except: %i[new edit] do
+      post :generate_metadata_suggestions, on: :collection
+      post :generate_metadata_suggestions, on: :member
       resources :decorations,
         only: %i[create update destroy],
         controller: :outfit_folder_decorations